- WriteOp: operator disposal, deletion of target files on cancelation - OperatorContext: constant for 'Operation canceled.' message - LookupTable etc.: API for float values
git-svn-id: https://www.brockmann-consult.de/svn/os/beam-4.x/trunk@2026 d90d2d48-2b33-0410-a973-e72f85796a02

+++ beam-core/src/main/java/org/esa/beam/util/math/LookupTable.java
@@ -127,7 +127,7 @@ public class LookupTable {
      * @throws NullPointerException     if the {@code values} array or the {@code dimensions} array
      *                                  is {@code null} or any dimension is {@code null}.
      */
-    public LookupTable(final float[] values, final double[]... dimensions) {
+    public LookupTable(final float[] values, final float[]... dimensions) {
         this(values, IntervalPartition.createArray(dimensions));
     }
